refactor: migrate color tool
