+++ packages/mask/src/utils/theme/index.ts
@@ -1,5 +1,4 @@
 // Do not export more files. Those files are deprecated.
-export * from './color-tools.js'
 export * from './MaskTheme.js'
 export * from './useThemeLanguage.js'
 export * from './useMaskSiteAdaptorMixedTheme.js'
