+++ packages/plugin-infra/src/entry-content-script.ts
@@ -45,3 +45,5 @@ export {
     useThemeSettings,
     useThemeSize,
 } from './sns-adaptor/index.js'
+
+export { getBackgroundColor, isDark, toRGB, fromRGB, shade } from './utils/theme/color-tools.js'
